add svn ignore properties to each module
Introduce concept of a component which is like a service but without a lifecycle.  Will use this as the base classes for everything that is currently a service and then create a proxy that wraps those objects to make them services.  This allows all the current service objects to be used outside the idP without pulling in the somewhat substantial overhead of service configuration and management.

Do a bit of work on the resolver