https://issues.shibboleth.net/jira/browse/IDP-25 Port SAML2 attribute encoder (and test) to V3 APIs.
Very simple given the support functions.  A bit of unsightly cut and paste, but without multiple inheritance what are you going to do? The V2 code has the same problem, but in V2 it was the SAML2 aspects that had to be duplicated, in V3 it is the Scoped bits which are duplicated and that is just setters and getters.