IDP-408 remove pointless synchronization
In general many of these can be removed - in particular any setter which has

ComponentSupport.ifInitializedThrowUnmodifiabledComponentException(this);

Need not be serialized.

Equally strategy setters (a very comon leitmotif, probably cut and pastes derived) have no reason to be synchronized;  If the use of the strategy has no synchronization there can be no value and assignment can be assumed to be atomic.

OTOH sereval serializers are syncrhonized and these need to stay as such - we do not want a class being changed during serialization.  These have been left untouched, as have all the syncrhonization for services.