IDP-804 Add 'impl' suffix to packages which are impl.
https://issues.shibboleth.net/jira/browse/IDP-804

This checkin does the (deprecated) RelyingParty and MetadataProvider syntax.  With the former being deprecated, everything goes into impl.  For the latter we have a hand full of factories (including one we accidentally documented).  Notably the only parser support we leave in place are the AbstractMetadataParser and the AbstractReloadingMetadataProviderParser.  The AbstractDynamic.....Parsers are left in impl.  We can always promote them (but never demote them)