IDP-804 Add 'impl' suffix to packages which are impl.
https://issues.shibboleth.net/jira/browse/IDP-804

This checkin does the Attribute Filter syntax.  This is important since much of the heavy lifting is done by the abstract classes for the Matcher and Policy.

This does in turn require that some constants are moved up from the impl classes to the BaseFilter since the work done by these abstract classes need to look at the QNames for the parents (which would usually be defined in their parsers.