IDP-645 Functional SLO for CAS protocol.
HTML5 sessionStorage is used for coordination between master logout
propagation flow UI and SP-specific flows. In the CAS case, no response
is received, so a Javascript-based mechanism is needed to trigger return
to the IdP to complete the propgagation flow. The sessionStorage facility
is a convenient state store between the logout-propagation UI and
individual iframes.